requirements.txt
Django==5.2.8
gunicorn==23.0.0
psycopg2-binary==2.9.10